fix(web): fix statusbar change app specific display
